+++ extensions/resource-deployment/src/ui/resourceTypePickerDialog.ts
@@ -10,11 +10,7 @@ import { IResourceTypeService } from '../services/resourceTypeService';
 import * as vscode from 'vscode';
 import { ResourceType, DeploymentProvider } from '../interfaces';
 import { IToolsService } from '../services/toolsService';
-import { INotebookService } from '../services/notebookService';
 import { DialogBase } from './dialogBase';
-import { NotebookInputDialog } from './deploymentDialog';
-import { IDownloadService } from '../services/downloadService';
-import * as cp from 'child_process';
 
 const localize = nls.loadMessageBundle();
 
@@ -29,10 +25,8 @@ export class ResourceTypePickerDialog extends DialogBase {
 	private _optionDropDownMap: Map<string, azdata.DropDownComponent> = new Map();
 
 	constructor(private extensionContext: vscode.ExtensionContext,
-		private notebookService: INotebookService,
 		private toolsService: IToolsService,
 		private resourceTypeService: IResourceTypeService,
-		private downloadService: IDownloadService,
 		resourceType: ResourceType) {
 		super(localize('resourceTypePickerDialog.title', "Select the deployment options"), 'ResourceTypePickerDialog', true);
 		this._selectedResourceType = resourceType;
@@ -182,32 +176,7 @@ export class ResourceTypePickerDialog extends DialogBase {
 	}
 
 	private onComplete(): void {
-		const self = this;
-		const provider = this.getCurrentProvider();
-		if (provider.dialog) {
-			const dialog = new NotebookInputDialog(this.notebookService, provider.dialog);
-			dialog.open();
-		} else if (provider.notebook) {
-			this.notebookService.launchNotebook(provider.notebook);
-		} else if (provider.executable) {
-			const taskName = localize('resourceDeployment.DownloadAndLaunchTaskName', "Download and launch installer");
-			azdata.tasks.startBackgroundOperation({
-				displayName: taskName,
-				description: taskName,
-				isCancelable: false,
-				operation: op => {
-					op.updateStatus(azdata.TaskStatus.InProgress, localize('resourceDeployment.DownloadingText', "Downloading from: {0}", provider.executable));
-					self.downloadService.download(provider.executable).then((downloadedFile) => {
-						op.updateStatus(azdata.TaskStatus.InProgress, localize('resourceDeployment.DownloadCompleteText', "Successfully downloaded: {0}", downloadedFile));
-						op.updateStatus(azdata.TaskStatus.InProgress, localize('resourceDeployment.LaunchingProgramText', "Launching: {0}", downloadedFile));
-						cp.exec(downloadedFile);
-						op.updateStatus(azdata.TaskStatus.Succeeded, localize('resourceDeployment.ProgramLaunchedText', "Successfully launched: {0}", downloadedFile));
-					});
-				}
-			});
-		} else if (provider.url) {
-			vscode.commands.executeCommand('vscode.open', vscode.Uri.parse(provider.url));
-		}
+		this.resourceTypeService.startDeployment(this.getCurrentProvider());
 		this.dispose();
 	}
 }
